requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
docopt==0.6.2
idna==2.9
multidict==4.7.5
pipreqs==0.4.10
python-dotenv==0.13.0
requests==2.23.0
urllib3==1.25.9
websockets==8.1
yarg==0.1.9
yarl==1.4.2